refactor(mempool-service): extract exclude-filter into subscriber submodule
Move-only extraction of TxIdExcludeSuffix, MempoolFilterError and
unique_suffix_match (with its tests) into subscriber/filter.rs. The
MempoolSubscriber filter methods stay put and call into the submodule.
No behavior, signature, or public-API change.
